Data preprocessing
In this file, the data has been preprocessed; Balanced, Scaled, Shuffled, and split into 3 subsets: training, validation, and test. Then the data has been saved as npz file. To load the data in TensorFlow, data needs to be in npz format. The actual model building will be done in another file which will be attached in this same repo soon.