Model building
The previous file has data preprocessed and was saved in a npz file format. In this file, the npz file data is loaded and a model is built and trained to predict whether a customer would return back or not. The hyper parameters like hidden layer size, number of hidden layers, activation functions, batch size, epochs can be changed to change the performance of the model.